Major: update functionality, docs, tests
Functionality: the core connect and query functions now work as I expect
them to, meaning that this is probably just about ready for a 0.1.0
release.

Docs: I've documented the functionality (such as it is) pretty
thoroughly in the README.

Tests: not a whole lot of coverage here, but it's a start. I'll get
around to mocking the PuppetDB calls soon, but until then there are some
decent unit tests.