Added query-params to GET and refactored query
Rather than manually tacking on "?query=..." to the path, the query
function now passes a map containing a :query key to the new and
improved GET function. The GET function takes the map, including
possibly additional query-string parameters, and forms the query-string
itself.

This doesn't change any functionality here, really, but it does lay the
groundwork for paging. I'm not crazy about the implementation here... I
wish that it were possible to pass the query map along to `query`
directly, but that's going to require some argument inspection (gotta
differentiate between query-vec and param-map).